+++ src/ViewModels/StatsShipsWaresTransactionsModel.cs
@@ -24,7 +24,7 @@ public sealed class StatsShipsWaresTransactionsModel : StatsShipsWaresBaseModel
   public StatsShipsWaresTransactionsModel()
     : base() { }
 
-  protected override List<(int ShipId, string ShipName, string WareId, string WareName, double Profit)> LoadData()
+  protected override List<(long ShipId, string ShipName, string WareId, string WareName, double Profit)> LoadData()
   {
     IEnumerable<Transaction> q = MainViewModel.AllTransactions;
     if (SelectedShipClass != "All")
